Run futurize over all code

+++ bluebottle/pages/admin.py
@@ -68,7 +68,7 @@ class PageAdmin(PlaceholderFieldAdmin):
         return urlpatterns + base_urls
 
     def get_base_object(self, pk):
-        pk = long(pk)
+        pk = int(pk)
         if pk:
             return Page.objects.get(pk=pk)
         else:
